REFACTOR: Decompose cmd/igor/main.go into single-responsibility package main files
